requirements.txt
sqlalchemy
flask
passlib
pymysql
python-dateutil
python-magic
flask-wtf
gunicorn
githubpy
requests
pyIsEmail
ipaddress
GitPython
xmltodict
lxml
pytz
tzlocal
libvirt-python; sys_platform != 'win32'